auth and client moved
