+++ app/src/main/java/com/m/services/appAuth/AuthStateManager.java
