+++ app/src/main/java/com/m/services/dataSync/Client.java
